Add wasm test
